+++ py/fiberassign/fba_launch_io.py
@@ -110,7 +110,7 @@ def get_program_latest_timestamp(
 
     Args:
         survey: survey (string; e.g. "sv1", "sv2", "sv3", "main")
-        program: ideally "dark", "bright", or "backup" (string)
+        program: ideally "dark", "dark1b", "bright", "bright1b", or "backup" (string)
                 though if different will return None
         tilera: tile center R.A. (float)
         tiledec: tile center Dec. (float)
@@ -130,6 +130,7 @@ def get_program_latest_timestamp(
     Notes:
         20210831: remove the +1min (and added "leq=True" in the read_targets_.. routines for ledgers)
         20220421: add too_tile optional argument
+        20250218: handle DARK1B/BRIGHT1B
     """
     # AR check DESI_SURVEYOPS is defined
     assert_env_vars(
@@ -143,11 +144,18 @@ def get_program_latest_timestamp(
     # AR check if the per-tile file is here
     # AR no need to check the scnd-mtl-done-tiles.ecsv file,
     # AR     as we restrict to a given program ([desi-survey 2434])
+    # AR DARK1B/BRIGHT1B: need to also parse DARK/BRIGHT
     fn = os.path.join(os.getenv("DESI_SURVEYOPS"), "mtl", "mtl-done-tiles.ecsv")
     if os.path.isfile(fn):
         d = Table.read(fn)
-        keep = d["PROGRAM"] == program.upper()
+        if program.upper() in ["DARK", "DARK1B"]:
+            keep = np.in1d(d["PROGRAM"], ["DARK", "DARK1B"])
+        elif program.upper in ["BRIGHT", "BRIGHT1B"]:
+            keep = np.in1d(d["PROGRAM"], ["BRIGHT", "BRIGHT1B"])
+        else:
+            keep = d["PROGRAM"] == program.upper()
         # AR add a cut on TILEID
+        # AR TODO: read tiles-{survey}.ecsv and use np.in1d()
         if survey == "sv3":
             keep &= (d["TILEID"] < 1000)
         if survey == "main":
@@ -169,8 +177,12 @@ def get_program_latest_timestamp(
         start=start,
     )
     # AR existing folders?
+    if isinstance(mtldir, list):
+        test_hpdirnames = mtldir + [scndmtldir]
+    else:
+        test_hpdirnames = [mtldir, scndmtldir]
     hpdirnames = []
-    for hpdirname in [mtldir, scndmtldir]:
+    for hpdirname in test_hpdirnames:
         if hpdirname is not None:
             if os.path.isdir(hpdirname):
                 hpdirnames.append(hpdirname)
@@ -733,7 +745,7 @@ def get_ledger_paths(
 
     Args:
         survey: survey (string; e.g. "sv1", "sv2", "sv3", "main")
-        program: "dark", "bright", or "backup" (string)
+        program: "dark", "dark1b", "bright", "bright1b" or "backup" (string)
         too_tile (optional, defaults to False):
                 if False and survey=="main" and ToO-fiber.ecsv exists, use ToO-fiber.ecsv (fiber-override only),
                 else, use ToO.ecsv (ToO for dedicated tiles)
@@ -744,18 +756,26 @@ def get_ledger_paths(
         start(optional, defaults to time()): start time for log (in seconds; output of time.time()
 
     Returns:
-        - mtl: primary ledgers MTL folder (string)
+        - mtl: primary ledgers MTL folder (string, or list of strings for "main" survey)
         - scndmtl: secondary ledgers MTL folder (string)
         - too: ToO ecsv ledger file (string)
 
     Notes:
         if survey not in ["sv1", "sv2", "sv3", "main"]
-        or program not in ["dark", "bright", or "backup"], will return a warning only
+        or program not in ["dark", "dark1b", "bright", "bright1b" or "backup"], will return a warning only
         same warning only if the built paths/files do not exist.
+        20250221: handle dark1b/bright1b:
+            - primary ledgers:
+                dark -> query dark only
+                dark1b -> query dark and dark1b
+            - secondary ledgers:
+                dark -> query dark only
+                dark1b -> query dark1b only
+        [same for bright, bright1b]
     """
     # AR expected survey, program?
     exp_surveys = ["sv1", "sv2", "sv3", "main"]
-    exp_programs = ["dark", "bright", "backup"]
+    exp_programs = ["dark", "dark1b", "bright", "bright1b", "backup"]
     if survey.lower() not in exp_surveys:
         log.warning(
             "{:.1f}s\t{}\tunexpected survey={} ({}; proceeding anyway)".format(
@@ -778,8 +798,14 @@ def get_ledger_paths(
     mtl = os.path.join(
         os.getenv("DESI_SURVEYOPS"), "mtl", survey.lower(), program.lower(),
     )
-    # AR secondary (dark, bright; no secondary for backup)
-    if program.lower() in ["dark", "bright"]:
+    # AR handle dark1b/bright1b
+    if (survey.lower() == "main") & (program.lower() in ["dark1b", "bright1b"]):
+        progshort = program.lower().replace("1b", "")
+        mtl = mtl.replace(program.lower(), progshort)
+        mtl2 = mtl.replace(progshort, "{}1b".format(progshort))
+        mtl = [mtl, mtl2]
+    # AR secondary (dark, dark1b, bright, bright1b; no secondary for backup)
+    if program.lower() in ["dark", "bright", "dark1b", "bright1b"]:
         scndmtl = os.path.join(
             os.getenv("DESI_SURVEYOPS"),
             "mtl",
@@ -809,14 +835,19 @@ def get_ledger_paths(
                 time() - start, step, name, path,
             )
         )
-        if path is not None:
-            if not os.path.exists(path):
-                log.warning(
-                    "{:.1f}s\t{}\tdirectory for {}: {} does not exist".format(
-                        time() - start, step, name, path,
+        if isinstance(path, list):
+            paths_to_check = path
+        else:
+            paths_to_check = [path]
+        for path_to_check in paths_to_check:
+            if path_to_check is not None:
+                if not os.path.exists(path_to_check):
+                    log.warning(
+                        "{:.1f}s\t{}\tdirectory for {}: {} does not exist".format(
+                            time() - start, step, name, path_to_check,
+                        )
                     )
-                )
-    #
+
     return mtl, scndmtl, too
 
 
@@ -839,7 +870,7 @@ def get_desitarget_paths(
     Args:
         dtver: desitarget catalog version (string; e.g., "0.57.0")
         survey: survey (string; e.g. "sv1", "sv2", "sv3", "main")
-        program: "dark", "bright", or "backup" (string)
+        program: "dark", "dark1b", "bright", "bright1b" or "backup" (string)
         too_tile (optional, defaults to False):
                 if False and survey=="main" and ToO-fiber.ecsv exists, use ToO-fiber.ecsv (fiber-override only),
                 else, use ToO.ecsv (ToO for dedicated tiles)
@@ -858,8 +889,8 @@ def get_desitarget_paths(
         - sky: sky folder
         - skysupp: skysupp folder
         - gfa: GFA folder
-        - targ: targets folder (static catalogs, with all columns)
-        - mtl: MTL folder
+        - targ: targets folder(s) (static catalogs, with all columns); 2-elt list for main/{dark1b,bright1b}, str otherwise
+        - mtl: MTL folder(s); 2-elt list for main/{dark1b,bright1b}, str otherwise
         - scnd: secondary fits catalog (static, with all columns)
         - scndmtl: MTL folder for secondary targets
         - scnd2, scnd3, etc: any other existing secondary fits catalog (static, with all columns)
@@ -867,17 +898,26 @@ def get_desitarget_paths(
 
     Notes:
         if survey not in ["sv1", "sv2", "sv3", "main"]
-        or program not in ["dark", "bright", or "backup"], will return a warning only
+        or program not in ["dark", "dark1b", "bright", "bright1b" or "backup"], will return a warning only
         same warning only if the built paths/files do not exist.
         20210917 : secondary -> add all existing folders (e.g., main2/) (backward-compatible change)
         20220318 : add custom_too_file optional argument
         20220421 : add too_tile optional argument
         20221004 : add custom_too_development optional argument
         20221031 : if multiple files provided in custom_too_file, then several keys: too, too2, too3, etc
+        20250221: handle dark1b/bright1b (targ2 and mtl2 keys are only present for survey=main, program=dark1b/bright1b)
+            - primary ledgers:
+                dark -> query dark only
+                dark1b -> query dark and dark1b
+            - secondary ledgers:
+                dark -> query dark only
+                dark1b -> query dark1b only
+        [same for bright, bright1b]
+
     """
     # AR expected survey, program?
     exp_surveys = ["sv1", "sv2", "sv3", "main"]
-    exp_programs = ["dark", "bright", "backup"]
+    exp_programs = ["dark", "dark1b", "bright", "bright1b", "backup"]
     if survey.lower() not in exp_surveys:
         log.warning(
             "{:.1f}s\t{}\tunexpected survey={} ({}; proceeding anyway)".format(
@@ -906,7 +946,7 @@ def get_desitarget_paths(
         dtcat = gaiadr
     else:
         dtcat = dr
-    mydirs["targ"] = os.path.join(
+    targ = os.path.join(
         os.getenv("DESI_TARGET"),
         "catalogs",
         dtcat,
@@ -916,8 +956,16 @@ def get_desitarget_paths(
         "resolve",
         program.lower(),
     )
-    # AR secondary (dark, bright; no secondary for backup)
-    if program.lower() in ["dark", "bright"]:
+    mydirs["targ"] = targ
+    if survey.lower() == "main":
+        progshort = program.lower().replace("1b", "")
+        if program.lower() == "{}1b".format(progshort):
+            targ = targ.replace(program.lower(), progshort)
+            targ2 = targ.replace(progshort, "{}1b".format(progshort))
+            mydirs["targ"] = [targ, targ2]
+    # AR secondary (dark, dark1b, bright, bright1b; no secondary for backup)
+    # AR only query program (in particular, only dark1b for dark1b; same for bright1b)
+    if program.lower() in ["dark", "bright", "dark1b", "bright1b"]:
         if survey.lower() == "main":
             basename = "targets-{}-secondary.fits".format(program.lower())
         else:
@@ -971,12 +1019,17 @@ def get_desitarget_paths(
                 time() - start, step, key, mydirs[key]
             )
         )
-        if not os.path.exists(mydirs[key]):
-            log.warning(
-                "{:.1f}s\t{}\tdirectory for {}: {} does not exist".format(
-                    time() - start, step, key, mydirs[key]
+        if isinstance(mydirs[key], list):
+            paths_to_check = mydirs[key]
+        else:
+            paths_to_check = [mydirs[key]]
+        for path_to_check in paths_to_check:
+            if not os.path.exists(path_to_check):
+                log.warning(
+                    "{:.1f}s\t{}\tdirectory for {}: {} does not exist".format(
+                        time() - start, step, key, path_to_check
+                    )
                 )
-            )
 
     # AR ledgers
     mydirs["mtl"], scndmtl, mydirs["too"] = get_ledger_paths(
@@ -1040,7 +1093,7 @@ def create_tile(
     tiledec,
     outfn,
     survey,
-    obscon="DARK|GRAY|BRIGHT|BACKUP",
+    obscon="DARK|DARK1B|GRAY|BRIGHT|BRIGHT1B|BACKUP",
     log=Logger.get(),
     step="",
     start=time(),
@@ -1054,7 +1107,7 @@ def create_tile(
         tiledec: tile center Dec. (float)
         outfn: fits file name to be written
         survey: survey (string; e.g. "sv1", "sv2", "sv3", "main")
-        obscon (optional, defaults to "DARK|GRAY|BRIGHT|BACKUP"): tile allowed observing conditions (string)
+        obscon (optional, defaults to "DARK|DARK1B|GRAY|BRIGHT|BRIGHT1B|BACKUP"): tile allowed observing conditions (string)
         log (optional, defaults to Logger.get()): Logger object
         step (optional, defaults to ""): corresponding step, for fba_launch log recording
             (e.g. dotiles, dosky, dogfa, domtl, doscnd, dotoo)
@@ -1310,7 +1363,7 @@ def create_mtl(
 
     Args:
         tilesfn: path to a tiles fits file (string)
-        mtldir: desisurveyops MTL folder (string)
+        mtldir: desisurveyops MTL folder (string or list of two strings)
         mtltime: MTL isodate (string formatted as yyyy-mm-ddThh:mm:ss+00:00)
         targdirs: desitarget targets folder (or file name(s) if secondary) for static fits catalog(s) (string or list)
         survey: survey (string; e.g. "sv1", "sv2", "sv3", "main")
@@ -1351,6 +1404,7 @@ def create_mtl(
                     - will exclude STD_WD as no rundate used
                     - guess the program from the mtldir
                     - only enabled for survey="main"
+        20250218 : handle desi-extension, with dark1b/bright1b
     """
     log.info("")
     log.info("")
@@ -1365,6 +1419,29 @@ def create_mtl(
     # AR    because parse_assign as imports from targets.py,
     # AR    which have imports from fba_launch_io.py...
     from fiberassign.assign import minimal_target_columns
+
+    # AR check on mtl
+    if isinstance(mtldir, list):
+        # AR only two ledgers allowed (e.g. dark, dark1b)
+        if len(mtldir) > 2:
+            msg = "len(mtldir) = {} > 2; only up two ledger folders can be provided".format(
+                len(mtldir)
+            )
+            log.error(msg)
+            raise ValueError(msg)
+        # AR sanity check: in that case, targdirs should also be a 2-elt list
+        if not isinstance(targdirs, list):
+            msg = "two ledger paths are provided ({}) but only one static path ({})".format(
+                ",".join(mtldir), targdirs
+            )
+            log.error(msg)
+            raise ValueError(msg)
+        if len(targdirs) > 2:
+            msg = "len(targdir) = {} > 2; only up two static folders can be provided".format(
+                len(targdir)
+            )
+            log.error(msg)
+            raise ValueError(msg)
 
     # AR change targdirs to list if a string is provided
     if isinstance(targdirs, str):
@@ -1381,6 +1458,9 @@ def create_mtl(
             isodate=mtltime,
         )
     else:
+        # AR read_targets_in_tiles() handles if mtldir is a single folder
+        # AR or a list with two folders
+        # AR https://github.com/desihub/desitarget/pull/833
         d = read_targets_in_tiles(
             mtldir,
             tiles=tiles,
@@ -1389,6 +1469,7 @@ def create_mtl(
             unique=True,
             isodate=mtltime,
             leq=True,
+            maketwostyle=True, # AR add MTL_{HIGHEST,WANTED,CONTAINS} columns
         )
     log.info(
         "{:.1f}s\t{}\treading {} targets from {}".format(
@@ -1456,9 +1537,66 @@ def create_mtl(
                 d, targdirs[0], columns=columns, header=False, strictcols=False, quick=True
             )
         else:
-            targ = read_targets_in_tiles(
-                targdirs[0], tiles=tiles, quick=True, columns=columns + ["TARGETID"]
+            targ = vstack(
+                [
+                    Table(
+                        read_targets_in_tiles(
+                            targdir, tiles=tiles, quick=True, columns=columns + ["TARGETID"]
+                        )
+                    ) for targdir in targdirs
+                ]
             )
+            log.info(
+                "{:.1f}s\t{}\tread {} rows from {}".format(
+                    time() - start, step, len(targ), ",".join(targdirs)
+                )
+            )
+            if len(targdirs) == 1:
+                ok = np.unique(targ["TARGETID"]).size == len(targ)
+                if not ok:
+                    msg = "found {} duplicates, but none are expected".format(
+                        len(targ) - np.unique(targ["TARGETID"]).size
+                    )
+                    log.error(msg)
+                    raise ValueError(msg)
+            else:
+                # AR remove possible duplicates using TARGETID
+                # AR for the static catalogs, all columns should be the same
+                # AR    e.g. in dark and dark1b for a given TARGETID
+                # AR    as TARGETID encodes the catalog release, etc
+                # AR so we do not perform any sanity check that duplicates
+                # AR    have the same column values
+                _, ii = np.unique(targ["TARGETID"], return_index=True)
+                # AR sanity check: dups should have MTL_CONTAINS corresponding to
+                # AR    both ledgers
+                _ = np.arange(len(targ), dtype=int)
+                dups_ii = _[~np.in1d(_, ii)]
+                dups_tids = np.unique(targ["TARGETID"][dups_ii])
+                oc0 = os.path.normpath(targdirs[0]).split(os.path.sep)[-1].upper()
+                oc1 = os.path.normpath(targdirs[1]).split(os.path.sep)[-1].upper()
+                val = obsconditions[oc0] | obsconditions[oc1]
+                expect_dups_ii = np.where(d["MTL_CONTAINS"] == val)[0]
+                expect_dups_tids = np.unique(d["TARGETID"][expect_dups_ii])
+                if not np.all(dups_tids == expect_dups_tids):
+                    msg = "discrepancy in duplicates from the static catalogs ({}) " \
+                        "vs. what is expected from the ledgers ({}), " \
+                        "ndiff = {}".format(
+                            dups_tids.size,
+                            expect_dups_tids.size,
+                            np.max([
+                                (~np.in1d(dups_ii, expect_dups_tids)).sum(),
+                                (~np.in1d(expect_dups_ii, dups_ii)).sum()
+                            ])
+                    )
+                    log.error(msg)
+                    raise ValueError(msg)
+                log.info(
+                    "{:.1f}s\t{}\tremove {} duplicates, remain {} rows".format(
+                        time() - start, step, len(targ) - ii.size, ii.size
+                    )
+                )
+                targ = targ[ii]
+            # AR now match to ledgers
             d = match_ledger_to_targets(d, targ)
     # AR mtl: case secondary
     else:
@@ -2117,7 +2255,8 @@ def update_fiberassign_header(
         sys.exit(1)
 
     # AR check on goaltype - faprgrm
-    if args.goaltype != args.program:
+    # AR (DARK1B->DARK, BRIGHT1B->BRIGHT)
+    if args.goaltype != args.program.replace("1B", ""):
         log.warning(
             "{:.1f}s\t{}\tgoaltype={} != program={}".format(
                 time() - start, step, args.goaltype, args.program,
@@ -2134,14 +2273,16 @@ def update_fiberassign_header(
     desiroot = os.getenv("DESI_ROOT")
     fd["PRIMARY"].write_key("DESIROOT", desiroot)
     for key in np.sort(list(mydirs.keys())):
-        if (key == "mtl") & (isinstance(mydirs["mtl"], list)):
-            # AR header keywords: MTL, MTL2, MTL3, etc
-            # AR probably to be deprecate for sv2
-            suffixs = [""] + np.arange(2, len(mydirs["mtl"]) + 1).astype(str).tolist()
-            for mtldir, suffix in zip(mydirs["mtl"], suffixs):
-                fd["PRIMARY"].write_key(
-                    "mtl{}".format(suffix), mtldir.replace(desiroot, "DESIROOT"),
-                )
+        if key in ["mtl", "targ"]:
+            if isinstance(mydirs[key], list):
+                # AR header keywords: MTL, MTL2 (or TARG, TARG2)
+                suffixs = [""] + np.arange(2, len(mydirs[key]) + 1).astype(str).tolist()
+                for path, suffix in zip(mydirs[key], suffixs):
+                    fd["PRIMARY"].write_key(
+                        "{}{}".format(key, suffix), path.replace(desiroot, "DESIROOT"),
+                    )
+            else:
+                fd["PRIMARY"].write_key(key, mydirs[key].replace(desiroot, "DESIROOT"))
         else:
             fd["PRIMARY"].write_key(key, mydirs[key].replace(desiroot, "DESIROOT"))
     # AR storing some specific arguments
@@ -2308,7 +2449,7 @@ def get_qa_tracers(
 
     Args:
         survey: survey name: "sv1", "sv2", "sv3" or "main") (string)
-        program: "DARK", "BRIGHT", or "BACKUP" (string)
+        program: "DARK", "BRIGHT", "DARK1B", "BRIGHT1B", or "BACKUP" (string)
         log (optional, defaults to None): Logger object
         step (optional, defaults to ""): corresponding step, for fba_launch log recording
             (e.g. dotiles, dosky, dogfa, domtl, doscnd, dotoo)
@@ -2321,7 +2462,10 @@ def get_qa_tracers(
     if program == "DARK":
         trmskkeys = ["DESI_TARGET", "DESI_TARGET", "DESI_TARGET"]
         trmsks = ["LRG", "ELG", "QSO"]
-    elif program == "BRIGHT":
+    elif program == "DARK1B":
+        trmskkeys = ["DESI_TARGET", "DESI_TARGET", "DESI_TARGET", "DESI_TARGET"]
+        trmsks = ["LRG", "LGE", "ELG", "QSO"]
+    elif program in ["BRIGHT", "BRIGHT1B"]:
         trmskkeys = ["BGS_TARGET", "BGS_TARGET"]
         trmsks = ["BGS_BRIGHT", "BGS_FAINT"]
         trmskkeys += ["MWS_TARGET", "MWS_TARGET"]
@@ -2333,15 +2477,14 @@ def get_qa_tracers(
         trmskkeys = ["MWS_TARGET", "MWS_TARGET", "MWS_TARGET"]
         trmsks = ["BACKUP_BRIGHT", "BACKUP_FAINT", "BACKUP_VERY_FAINT"]
     else:
+        msg = "{:.1f}s\t{}\tprogram={} not in DARK, BRIGHT, DARK1B, BRIGHT1B, or BACKUP; exiting".format(
+            time() - start, step, program
+        )
         if log is not None:
-            log.error(
-                "{:.1f}s\t{}\tprogram={} not in DARK, BRIGHT, or BACKUP; exiting".format(
-                    time() - start, step, program
-                )
-            )
+            log.error(msg)
         else:
-            print("program={} not in DARK, BRIGHT, or BACKUP; exiting".format(program))
-        sys.exit(1)
+            print(msg)
+        raise ValueError(msg)
 
     return trmskkeys, trmsks
 
@@ -2602,7 +2745,7 @@ def qa_print_infos(
     Args:
         ax: pyplot object
         survey: "sv1", "sv2", "sv3" or "main" (string)
-        program: "DARK", "BRIGHT", or "BACKUP" (string)
+        program: "DARK", "BRIGHT", "DARK1B", "BRIGHT1B", or "BACKUP" (string)
         faflavor: usually {survey}{program} in lower cases (string)
         tileid: tile TILEID (int)
         tilera: tile center R.A. in degrees (float)
@@ -2612,6 +2755,8 @@ def qa_print_infos(
         parent: dictionary for the parent target sample (output by get_parent_assign_quants())
         assign: dictionary for the assigned target sample (output by get_parent_assign_quants())
     """
+    assert program in ["DARK", "BRIGHT", "DARK1B", "BRIGHT1B", "BACKUP"]
+
     # AR hard-setting the plotted tracers
     # AR TBD: handle secondaries
     trmskkeys, trmsks = get_qa_tracers(survey, program)
@@ -2661,9 +2806,9 @@ def qa_print_infos(
     # AR infos: brightest target and assigned object
     # AR infos: taking a default 16, in case new programs are added
     magthresh = 16.0
-    if program == "DARK":
+    if program in ["DARK", "DARK1B"]:
         magthres = 16.0
-    if program == "BRIGHT":
+    if program in ["BRIGHT", "BRIGHT1B"]:
         magthresh = 15.0
     if program == "BACKUP":
         magthresh = 15.0
@@ -3099,6 +3244,8 @@ def plot_hist_tracer(ax, survey, parent, assign, msk, mskkey):
             band = "R"
         if "LRG" in msk:
             band = "Z"
+        if "LGE" in msk:
+            band = "Z"
         if "ELG" in msk:
             band = "G"
         if "QSO" in msk:
@@ -3418,17 +3565,19 @@ def make_qa(
     Args:
         outpng: written output PNG file (string)
         survey: "sv1", "sv2", "sv3" or "main" (string)
-        program: "DARK", "BRIGHT", or "BACKUP" (string)
+        program: "DARK", "BRIGHT", "DARK1B", "BRIGHT1B", or "BACKUP" (string)
         faflavor: usually {survey}{program} in lower cases (string)
         fiberassignfn: path to the output fiberassign-TILEID.fits file (string)
         tileid: tile TILEID (int)
         tilera: tile center R.A. in degrees (float)
         tiledec: tile center Dec. in degrees (float)
-        obscon: tile allowed observing conditions (string; e.g. "DARK|GRAY|BRIGHT|BACKUP")
+        obscon: tile allowed observing conditions (string; e.g. "DARK|DARK1B|GRAY|BRIGHT|BRIGHT1B||BACKUP")
         rundate: used rundate (string)
         tmpoutdir (optional, defaults to a temporary directory): temporary directory (to download the cutout)
         width_deg (optional, defaults to 4): width of the cutout in degrees (np.array of floats)
     """
+    assert program in ["DARK", "BRIGHT", "DARK1B", "BRIGHT1B", "BACKUP"]
+
     # AR WD and STD used masks
     _, wd_mskkeys, wd_msks, std_mskkeys, std_msks = get_dt_masks(survey, rundate=rundate)