+++ doc/changes.rst
@@ -6,6 +6,7 @@ fiberassign change log
 5.8.0 (unreleased)
 ------------------
 
+* handle dark1b/bright1b (PR `#468`_)
 * Custom too: allow ToO from fiberassign tertiary folder (PR `#476`_)
 * add stuck_on_sky_from_fafns() option (PR `#471`_)
 * get_fba_use_fabs(): add log infos (PR `#478`_)
@@ -16,6 +17,7 @@ fiberassign change log
 * Fix C++ bug with use of std::abs and improve floating point reproducibility (PR `#470`_).
 * Remove ``DesiTest`` from setup.py and warn about other deprecated features (PR `#464`_).
 
+.. _`#468`: https://github.com/desihub/fiberassign/pull/468
 .. _`#476`: https://github.com/desihub/fiberassign/pull/476
 .. _`#471`: https://github.com/desihub/fiberassign/pull/471
 .. _`#478`: https://github.com/desihub/fiberassign/pull/478